requirements.txt
asgiref==3.4.1
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.9
cssselect==1.1.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.2.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-import-export==2.7.1
django-js-asset==1.2.2
django-money==2.1
django-userforeignkey==0.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.0.0
docx==0.2.4
et-xmlfile==1.1.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
importlib-metadata==4.8.2
lxml==4.6.4
Markdown==3.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
psycopg2==2.9.3
psycopg2-binary==2.9.1
py-moneyed==1.2
Pygments==2.10.0
PyJWT==2.3.0
pyquery==1.4.3
python-dateutil==1.5
python-decouple==3.6
python-docx==0.8.11
pytz==2021.3
PyYAML==5.3.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.3.1
tablib==3.1.0
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.1.9
wfastcgi==3.0.0
whitenoise==5.2.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.6.0
